requirements.txt
fastapi
uvicorn
alembic
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
pyjwt